Tests/LibWeb/Text/input/HTML/Window-find-clear-selection.html

<!DOCTYPE html>
<script src="../include.js"></script>
<div>test</div>
<script>
    test(() => {
        let result = window.find("test");
        println(`window.find("test") initial result: ${result}`);
        window.getSelection().removeAllRanges();
        result = window.find("test");
        println(`window.find("test") after clearing the previous selection: ${result}`);
    });
</script>
