Tests/LibWeb/Text/expected/HTML/Window-find-clear-selection.txt

window.find("test") initial result: true
window.find("test") after clearing the previous selection: true
